requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0
pydantic-settings>=2.0
faster-whisper>=1.1.0
httpx>=0.27.0
openai>=1.50.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0
